move to_XXX curve methods to core.py
