+++ desc/geometry/curve.py
@@ -199,70 +199,6 @@ class FourierRZCurve(Curve):
                 + f"basis with {self.Z_basis.num_modes} modes"
             )
 
-    def to_FourierXYZCurve(self, N=None, grid=None, phis=None, name=""):
-        """Convert Curve to FourierXYZCurve representation.
-
-        Parameters
-        ----------
-        N : int
-            Fourier resolution of the new X,Y,Z representation.
-        grid : Grid, int or None
-            Grid used to evaluate curve coordinates on to fit with FourierXYZCurve.
-            If an integer, uses that many equally spaced points.
-        phis : ndarray
-            angle to use for the fitting. if None, defaults to
-            normalized arclength
-        name : str
-            name for this curve
-
-        Returns
-        -------
-        curve : FourierXYZCurve
-            New representation of the curve parameterized by Fourier series for X,Y,Z.
-
-        """
-        coords = self.compute("x", grid=grid, basis="xyz")["x"]
-        return FourierXYZCurve.from_values(
-            coords, N=N, phis=phis, basis="xyz", name=name
-        )
-
-    def to_SplineXYZCurve(self, knots=None, grid=None, method="cubic", name=""):
-        """Convert Curve to SplineXYZCurve.
-
-        Parameters
-        ----------
-        knots : ndarray
-            arbitrary theta values to use for spline knots,
-            should be an 1D ndarray of same length as the input.
-            (input length in this case is determined by grid argument, since
-            the input coordinates come from
-            Curve.compute("x",grid=grid))
-            If None, defaults to using an equal-arclength angle as the knots
-            If supplied, will be rescaled to lie in [0,2pi]
-        grid : Grid, int or None
-            Grid used to evaluate curve coordinates on to fit with SplineXYZCurve.
-            If an integer, uses that many equally spaced points.
-        method : str
-            method of interpolation
-            - `'nearest'`: nearest neighbor interpolation
-            - `'linear'`: linear interpolation
-            - `'cubic'`: C1 cubic splines (aka local splines)
-            - `'cubic2'`: C2 cubic splines (aka natural splines)
-            - `'catmull-rom'`: C1 cubic centripetal "tension" splines
-        name : str
-            name for this curve
-
-        Returns
-        -------
-        SplineXYZCurve: SplineXYZCurve
-            New representation of the curve parameterized by a spline for X,Y,Z.
-
-        """
-        coords = self.compute("x", grid=grid, basis="xyz")["x"]
-        return SplineXYZCurve.from_values(
-            coords, knots=knots, method=method, name=name, basis="xyz"
-        )
-
 
 class FourierXYZCurve(Curve):
     """Curve parameterized by Fourier series for X,Y,Z in terms of arbitrary angle phi.
@@ -483,46 +419,6 @@ class FourierXYZCurve(Curve):
         Z_n = transform.fit(coords_xyz[:, 2])
         return cls(X_n=X_n, Y_n=Y_n, Z_n=Z_n, name=name)
 
-    def to_SplineXYZCurve(self, knots=None, grid=None, method="cubic", name=""):
-        """Convert Curve to SplineXYZCurve.
-
-        Parameters
-        ----------
-        knots : ndarray
-            arbitrary theta values to use for spline knots,
-            should be an 1D ndarray of same length as the input.
-            (input length in this case is determined by grid argument, since
-            the input coordinates come from
-            Curve.compute("x",grid=grid))
-            If None, defaults to using an equal-arclength angle as the knots
-            If supplied, will be rescaled to lie in [0,2pi]
-        grid : Grid, int or None
-            Grid used to evaluate curve coordinates on to fit with SplineXYZCurve.
-            If an integer, uses that many equally spaced points.
-        method : str
-            method of interpolation
-            - `'nearest'`: nearest neighbor interpolation
-            - `'linear'`: linear interpolation
-            - `'cubic'`: C1 cubic splines (aka local splines)
-            - `'cubic2'`: C2 cubic splines (aka natural splines)
-            - `'catmull-rom'`: C1 cubic centripetal "tension" splines
-        name : str
-            name for this curve
-
-        Returns
-        -------
-        SplineXYZCurve: SplineXYZCurve
-            New representation of the curve parameterized by a spline for X,Y,Z.
-
-        """
-        coords = self.compute("x", grid=grid, basis="xyz")["x"]
-        return SplineXYZCurve.from_values(
-            coords, knots=knots, method=method, name=name, basis="xyz"
-        )
-
-    # TODO: to_rz method for converting to FourierRZCurve representation
-    # (might be impossible to parameterize with toroidal angle phi)
-
 
 class FourierPlanarCurve(Curve):
     """Curve that lies in a plane.
@@ -656,65 +552,6 @@ class FourierPlanarCurve(Curve):
             idx = self.r_basis.get_idx(0, 0, nn)
             if rr is not None:
                 self.r_n = put(self.r_n, idx, rr)
-
-    def to_FourierXYZCurve(self, N=None, grid=None, name=""):
-        """Convert Curve to FourierXYZCurve representation.
-
-        Parameters
-        ----------
-        N : int
-            Fourier resolution of the new X,Y,Z representation.
-        grid : Grid, int or None
-            Grid used to evaluate curve coordinates on to fit with FourierXYZCurve.
-            If an integer, uses that many equally spaced points.
-        name : str
-            name for this curve
-
-        Returns
-        -------
-        curve : FourierXYZCurve
-            New representation of the curve parameterized by Fourier series for X,Y,Z.
-
-        """
-        coords = self.compute("x", grid=grid, basis="xyz")["x"]
-        return FourierXYZCurve.from_values(coords, N=N, basis="xyz", name=name)
-
-    def to_SplineXYZCurve(self, knots=None, grid=None, method="cubic", name=""):
-        """Convert Curve to SplineXYZCurve.
-
-        Parameters
-        ----------
-        knots : ndarray
-            arbitrary theta values to use for spline knots,
-            should be an 1D ndarray of same length as the input.
-            (input length in this case is determined by grid argument, since
-            the input coordinates come from
-            Curve.compute("x",grid=grid))
-            If None, defaults to using an equal-arclength angle as the knots
-            If supplied, will be rescaled to lie in [0,2pi]
-        grid : Grid, int or None
-            Grid used to evaluate curve coordinates on to fit with SplineXYZCurve.
-            If an integer, uses that many equally spaced points.
-        method : str
-            method of interpolation
-            - `'nearest'`: nearest neighbor interpolation
-            - `'linear'`: linear interpolation
-            - `'cubic'`: C1 cubic splines (aka local splines)
-            - `'cubic2'`: C2 cubic splines (aka natural splines)
-            - `'catmull-rom'`: C1 cubic centripetal "tension" splines
-        name : str
-            name for this curve
-
-        Returns
-        -------
-        SplineXYZCurve: SplineXYZCurve
-            New representation of the curve parameterized by a spline for X,Y,Z.
-
-        """
-        coords = self.compute("x", grid=grid, basis="xyz")["x"]
-        return SplineXYZCurve.from_values(
-            coords, knots=knots, method=method, name=name, basis="xyz"
-        )
 
 
 class SplineXYZCurve(Curve):
@@ -873,25 +710,3 @@ class SplineXYZCurve(Curve):
         if basis == "rpz":
             coords = rpz2xyz(coords)
         return cls(coords[:, 0], coords[:, 1], coords[:, 2], knots, method, name)
-
-    def to_FourierXYZCurve(self, N=None, grid=None, name=""):
-        """Convert Curve to FourierXYZCurve representation.
-
-        Parameters
-        ----------
-        N : int
-            Fourier resolution of the new X,Y,Z representation.
-        grid : Grid, int or None
-            Grid used to evaluate curve coordinates on to fit with FourierXYZCurve.
-            If an integer, uses that many equally spaced points.
-        name : str
-            name for this curve
-
-        Returns
-        -------
-        curve : FourierXYZCurve
-            New representation of the curve parameterized by Fourier series for X,Y,Z.
-
-        """
-        coords = self.compute("x", grid=grid, basis="xyz")["x"]
-        return FourierXYZCurve.from_values(coords, N=N, basis="xyz", name=name)
